fix:[조승훈]이름 에러난 부분 제거 및 Response 추가 작성[PTANP-T1-X]

+++ answerProcess/product/service/response/ProductReadResponse.py
@@ -5,11 +5,11 @@ from dataclasses import dataclass
 class ProductReadResponse:
     __productNumber: int
     __productTitle: str
-    __productPrice: float
+    __productPrice: int
     __productDetails: str
     __seller: str
 
-    def __init__(self,productNumber: int, productTitle: str, productDetails: str, seller: str, productPrice: float):
+    def __init__(self,productNumber: int, productTitle: str, productDetails: str, seller: str, productPrice: int):
         self.__productNumber = productNumber
         self.__productTitle = productTitle
         self.__productDetails = productDetails
